castcast

+++ Mid/odometry.c
@@ -17,8 +17,8 @@ void get_crntCoordinate(struct coordinate* crnt_coordinate){
     float delta_L = 0.0; // 移動量
 
     /* 左右モータの回転量を計算する[rad] */
-    float delta_PhL = PI_FLOAT * (float)(motor_get_counts(left_motor)  - motor_get_pre_counts(left_motor))  / 180.0;
-    float delta_PhR = PI_FLOAT * (float)(motor_get_counts(right_motor) - motor_get_pre_counts(right_motor)) / 180.0;
+    float delta_PhL = PI_FLOAT * (float)(motor_get_counts(left_motor)  - motor_get_pre_counts(left_motor))  / 180.0f;
+    float delta_PhR = PI_FLOAT * (float)(motor_get_counts(right_motor) - motor_get_pre_counts(right_motor)) / 180.0f;
     // printf("delta_PhL:%f | delta_PhR:%f | ",delta_PhL,delta_PhR);
     
     /* 左右モータの移動量を計算する */
@@ -26,21 +26,21 @@ void get_crntCoordinate(struct coordinate* crnt_coordinate){
     float delta_LR = (float)(wheel_size / 2) * delta_PhR;
 
     // 走行体の旋回角度を計算する(エンコーダー)
-    float delta_rad_e = 0.0;
-    float delta_theta_e = 0.0;
+    float delta_rad_e = 0.0f;
+    float delta_theta_e = 0.0f;
 
-    if(delta_LL > 0.0 && delta_LR < 0.0){ //時計回りに旋回している
+    if(delta_LL > 0.0f && delta_LR < 0.0f){ //時計回りに旋回している
         //ロボットの移動距離
         delta_L = 0.0;
         // 旋回量の計算
         delta_rad_e = (abs(delta_LL) + abs(delta_LR)) / (float)wheel_dist;
 
     }
-    else if((delta_LL < 0.0 && delta_LR > 0.0)){ // 反時計回りに旋回している
+    else if((delta_LL < 0.0f && delta_LR > 0.0f)){ // 反時計回りに旋回している
         //ロボットの移動距離
-        delta_L = 0.0;
+        delta_L = 0.0f;
         // 旋回量の計算
-        delta_rad_e = -1 * (abs(delta_LL) + abs(delta_LR)) / (float)wheel_dist;
+        delta_rad_e = -1.0f * (abs(delta_LL) + abs(delta_LR)) / (float)wheel_dist;
 
     }
     else { //直進・曲進している
@@ -50,14 +50,14 @@ void get_crntCoordinate(struct coordinate* crnt_coordinate){
         delta_rad_e = (delta_LL - delta_LR) / (float)wheel_dist;
         
         if(abs(delta_rad_e) > 0.174){ //delta_rad_eが十分大きい
-            delta_L = 2 * (delta_L / delta_rad_e) * sin(delta_rad_e / 2);
+            delta_L = 2.0f * (delta_L / delta_rad_e) * (float)sin((double)(delta_rad_e / 2.0f));
         }
         
     }
-    delta_theta_e = (float)((delta_rad_e * 180.0) / PI_FLOAT);
+    delta_theta_e = (float)((delta_rad_e * 180.0f) / PI_FLOAT);
 
     /* 走行体の旋回角度を計算する (ジャイロ)*/
-    float delta_theta_g = gyro_sensor_get_angle(gyro_sensor) - gyro_sensor_get_pre_angle(gyro_sensor);
+    float delta_theta_g = (float)(gyro_sensor_get_angle(gyro_sensor) - gyro_sensor_get_pre_angle(gyro_sensor));
     float delta_rad_g   = PI_FLOAT * delta_theta_g /180.0;
 
     // 現在座標を計算する
@@ -65,8 +65,8 @@ void get_crntCoordinate(struct coordinate* crnt_coordinate){
     float delta_rad = delta_rad_e;      // 旋回量計算に使用するデバイスの選択
     float delta_theta = delta_theta_e;  // 旋回量計算に使用するデバイスの選択
 
-    crnt_coordinate->x      = pre_coordinate.x + (float)((double)delta_L * cos(pre_rad + (delta_rad / 2.0)));
-    crnt_coordinate->y      = pre_coordinate.y + (float)((double)delta_L * sin(pre_rad + (delta_rad / 2.0)));
+    crnt_coordinate->x      = pre_coordinate.x + (float)( (double)delta_L * cos((double)(pre_rad + (delta_rad / 2.0))) );
+    crnt_coordinate->y      = pre_coordinate.y + (float)( (double)delta_L * sin((double)(pre_rad + (delta_rad / 2.0))) );
     crnt_coordinate->theta  = pre_coordinate.theta + delta_theta;
     
     //前回座標を更新する
